+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.1.0
 ## Added
 * Android 11 (R) api version
+* Ability to generate just the diff with the `diff` command line option
 
 ## Fixed
 * signature generation in `ApiInfoStoreImpl`